.history/requirements_20241118005009.txt
openai==1.12.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
tqdm==4.66.1
torch==2.1.0
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99